requirements.txt
# ===== Core Framework =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ===== Database =====
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0

# ===== Redis & Celery =====
redis==5.0.1
celery[redis]==5.3.6
flower==2.0.1

# ===== HTTP & Scraping =====
httpx==0.26.0
aiohttp==3.9.1
feedparser==6.0.10
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ===== AI & Embeddings =====
anthropic==0.18.1
openai==1.12.0
tiktoken==0.5.2

# ===== Notifications =====
sendgrid==6.11.0
twilio==8.12.0

# ===== WebSocket =====
python-socketio==5.10.0

# ===== Authentication =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ===== Utilities =====
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
orjson==3.9.12

# ===== Testing =====
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-timeout==2.2.0
aiosqlite==0.19.0
fakeredis==2.20.1

# ===== Development =====
black==24.1.1
ruff==0.1.14
mypy==1.8.0
